v1.07, bug fixes, credits added, etc.
Added a credits button, in addition to making the HowToPlay button just be an animation to avoid issues with Scene switching causing bugs for the Host and Client. This should be a stable build, may add more, but not sure if that will be in the presentation or not. For now, this will do.